test(ping): Change test data structure to array
The test data structure to perform multiple request was a map. The map
in golang doesn't guarantee any order when its iterated through.

This cause flaky RateLimit tests because the order of the
requests is important to trigger the rate limit and to test the proper
response.

To fix that just use an array instead of a map. The array is an ordered
data structure.