feat(session): Add persistent session cookie
Enable nginx ingress persistent session cookie to enable
sticky pod session affinity.

So that all further requests from the same client goes to
the same pod instance. This way the rate limit can be
maintain in local memory and don't need a shared
storage.

Its not a 100 percent solution but in terms of effort and
performance its a real good simple one.