dependency(redis + dockertest):  Add redis client library
Add the most common redis client library for golang  [go-redis](https://github.com/go-redis/).

The dockertest library is used in the redis_bucket_test.go
integration test to spin up a docker redis container as part
of the go test itself. This go test has the tag integration
because it is not a unit test and for it to succeed it needs
a installed docker daemon.